ShaneGPUCountMinSketch

Documentation for ShaneGPUCountMinSketch.

ShaneGPUCountMinSketch.gpu_cms

ShaneGPUCountMinSketch.record

ShaneGPUCountMinSketch.check_and_fill_placeholder!

ShaneGPUCountMinSketch.count!

ShaneGPUCountMinSketch.filter_empty_seq!

ShaneGPUCountMinSketch.generate_combinations

ShaneGPUCountMinSketch.get_A_and_combs!

ShaneGPUCountMinSketch.get_max_nz_len

ShaneGPUCountMinSketch.gpu_cms — Type

R stores the random number for computing the hash Sk stores the sketch

d: the dimension of the configuration e.g. for a configuration of 3 elements, d = 3 e.g. (f₁, d₁₂, f₂) config of length 3 (f₁, d₁₂, f₂, d₂₃, f₃) config of length 5

ShaneGPUCountMinSketch.record — Type

Take non-zero code components i.e. Dict{Int => Vector{CartesianIndex{2}}} seq => [(pos, fil)] and convert it to an Array of records

nz_dict: non-zero code components
    Dict{Int => Vector{CartesianIndex{2}}}
    seq => [(pos, fil)]
num_fils: the number of filters in the configuration
# e.g. (f₁, d₁₂, f₂, d₂₃, f₃) has 3 filters

ShaneGPUCountMinSketch.check_and_fill_placeholder! — Method

obtain placeholder_count

ShaneGPUCountMinSketch.count! — Method

combs_gpu: the combinations of the non-zero code components
c: the CountMinSketch object
A: Array that contains the code
    A[:,1,n] contains the placements
    A[:,2,n] contains the corresponding filter indices
fil_len: the length of the filter; supply by the user

modify the CountMinSketch object c in place

note: A could be a parition (TODO)

ShaneGPUCountMinSketch.filter_empty_seq! — Method

filteremptyseq!(nz_dict::Dict{Int, Vector{CartesianIndex{2}}}) Filter out keys (seq) that have empty values

ShaneGPUCountMinSketch.generate_combinations — Method

input: numfils: the number of filters in the configuration maxnz_len: the maximum number of the non-zero code components (this needs to be counted first as to reduce the number of combinations)

ShaneGPUCountMinSketch.get_A_and_combs! — Method

getA!(nzdict::Dict{Int, Vector{CartesianIndex{2}}}) Take non-zero code components i.e. Dict{Int => Vector{CartesianIndex{2}}} seq => [(pos, fil)] and convert it to an Array of records

In another word: construct an Array A: Array that contains the code A[:,1,n] contains the placements A[:,2,n] contains the corresponding filter indices

In the process it also deletes the empty seqs in nz_dict

Note: can partition nz_dict into the 1st 1000 entries, 2nd 1000 entries, etc... in case the memory usage is a concern

returns the cpu and gpu version of A i.e. returns Acpu, Agpu

ShaneGPUCountMinSketch.get_max_nz_len — Method

Get maximum number of non-zero code components in our records This is to construct the upperbound of the Array of records